+++ src/main/java/com/science/gtnl/common/machine/multiblock/LargeBrewer.java
@@ -158,7 +158,7 @@ public class LargeBrewer extends GTMMultiMachineBase<LargeBrewer> implements ISu
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 3);
+        this.machineMode = (this.machineMode + 1) % 3;
         GTUtility.sendChatToPlayer(aPlayer, StatCollector.translateToLocal("LargeBrewer_Mode_" + this.machineMode));
     }
