+++ src/main/java/com/science/gtnl/common/machine/multiblock/TeleportationArrayToAlfheim.java
@@ -158,7 +158,7 @@ public class TeleportationArrayToAlfheim extends MultiMachineBase<TeleportationA
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 4);
+        this.machineMode = (this.machineMode + 1) % 4;
         GTUtility.sendChatToPlayer(
             aPlayer,
             StatCollector.translateToLocal("TeleportationArrayToAlfheim_Mode_" + this.machineMode));
