+++ src/main/java/com/science/gtnl/common/machine/multiblock/structuralReconstructionPlan/VacuumDryingFurnace.java
@@ -181,7 +181,7 @@ public class VacuumDryingFurnace extends GTMMultiMachineBase<VacuumDryingFurnace
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 4);
+        this.machineMode = (this.machineMode + 1) % 4;
         GTUtility
             .sendChatToPlayer(aPlayer, StatCollector.translateToLocal("VacuumDryingFurnace_Mode_" + this.machineMode));
     }
