+++ src/main/java/com/science/gtnl/common/machine/multiblock/wireless/CompoundExtremeCoolingUnit.java
@@ -255,7 +255,7 @@ public class CompoundExtremeCoolingUnit extends WirelessEnergyMultiMachineBase<C
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 3);
+        this.machineMode = (this.machineMode + 1) % 3;
         GTUtility.sendChatToPlayer(
             aPlayer,
             StatCollector.translateToLocal("CompoundExtremeCoolingUnit_Mode_" + this.machineMode));
