+++ src/main/java/com/science/gtnl/common/machine/multiblock/wireless/EngravingLaserPlant.java
@@ -306,7 +306,7 @@ public class EngravingLaserPlant extends WirelessEnergyMultiMachineBase<Engravin
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 2);
+        this.machineMode = (this.machineMode + 1) % 2;
         GTUtility
             .sendChatToPlayer(aPlayer, StatCollector.translateToLocal("EngravingLaserPlant_Mode_" + this.machineMode));
     }
