+++ src/main/java/com/science/gtnl/common/machine/multiblock/wireless/MegaCanner.java
@@ -195,7 +195,7 @@ public class MegaCanner extends WirelessEnergyMultiMachineBase<MegaCanner> {
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 2);
+        this.machineMode = (this.machineMode + 1) % 2;
         GTUtility.sendChatToPlayer(aPlayer, StatCollector.translateToLocal("MegaCanning_Mode_" + this.machineMode));
     }
