+++ src/main/java/com/science/gtnl/common/machine/multiblock/wireless/SuperconductingElectromagnetism.java
@@ -244,7 +244,7 @@ public class SuperconductingElectromagnetism extends WirelessEnergyMultiMachineB
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 3);
+        this.machineMode = (this.machineMode + 1) % 3;
         GTUtility.sendChatToPlayer(
             aPlayer,
             StatCollector.translateToLocal("SuperconductingElectromagnetism_Mode_" + this.machineMode));
